rag-app/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pgvector
sentence-transformers
docling
markdown-it-py
numpy